+++ src/polyfem/CMakeLists.txt
@@ -15,6 +15,7 @@ add_subdirectory(state)
 add_subdirectory(assembler)
 add_subdirectory(autogen)
 add_subdirectory(basis)
+add_subdirectory(io)
 add_subdirectory(mesh)
 add_subdirectory(problem)
 add_subdirectory(quadrature)
