+++ src/polyfem/io/CMakeLists.txt
@@ -1,0 +1,18 @@
+set(SOURCES
+	FEBioReader.cpp
+	FEBioReader.hpp
+	MatrixIO.cpp
+	MatrixIO.hpp
+	MshReader.cpp
+	MshReader.hpp
+	OBJReader.cpp
+	OBJReader.hpp
+	OBJWriter.cpp
+	OBJWriter.hpp
+	VTUWriter.cpp
+	VTUWriter.hpp
+)
+
+prepend_current_path(SOURCES)
+polyfem_set_source_group(${SOURCES})
+target_sources(polyfem PRIVATE ${SOURCES})
