+++ src/polyfem/io/FEBioReader.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <polyfem/State.hpp>
+
+#include <string>
+
+namespace polyfem::io
+{
+	class FEBioReader
+	{
+	public:
+		static void load(const std::string &path, const json &args_in, State &state, const std::string &export_solution = "");
+	};
+} // namespace polyfem::io
