+++ src/polyfem/mesh/mesh2D/CMesh2D.cpp
@@ -3,10 +3,8 @@
 
 #include <polyfem/mesh/MeshUtils.hpp>
 #include <polyfem/mesh/mesh2D/Refinement.hpp>
-
+#include <polyfem/io/MshReader.hpp>
 #include <polyfem/utils/StringUtils.hpp>
-#include <polyfem/utils/MshReader.hpp>
-
 #include <polyfem/utils/Logger.hpp>
 
 #include <geogram/basic/file_system.h>
@@ -19,6 +17,7 @@
 
 namespace polyfem
 {
+	using namespace io;
 	using namespace utils;
 
 	namespace mesh
@@ -50,7 +49,7 @@ namespace polyfem
 
 				mesh_.clear(false, false);
 
-				//TODO add tags to the refinement
+				// TODO add tags to the refinement
 				if (all_simplicial)
 				{
 					refine_triangle_mesh(mesh, mesh_);
@@ -96,7 +95,7 @@ namespace polyfem
 
 		bool CMesh2D::load(const std::string &path)
 		{
-			//This method should be used for special loading, like hybrid in 3d
+			// This method should be used for special loading, like hybrid in 3d
 
 			// edge_nodes_.clear();
 			// face_nodes_.clear();
@@ -211,7 +210,7 @@ namespace polyfem
 					orders_(f) = 1;
 					continue;
 				}
-				//P2
+				// P2
 				else if (nodes_ids.size() == 6)
 				{
 					orders_(f) = 2;
@@ -220,7 +219,7 @@ namespace polyfem
 					{
 						auto &n = edge_nodes_[index.edge];
 
-						//nodes not aleardy created
+						// nodes not aleardy created
 						if (n.nodes.size() <= 0)
 						{
 							n.v1 = index.vertex;
@@ -240,7 +239,7 @@ namespace polyfem
 						index = next_around_face(index);
 					}
 				}
-				//P3
+				// P3
 				else if (nodes_ids.size() == 10)
 				{
 					orders_(f) = 3;
@@ -249,7 +248,7 @@ namespace polyfem
 					{
 						auto &n = edge_nodes_[index.edge];
 
-						//nodes not aleardy created
+						// nodes not aleardy created
 						if (n.nodes.size() <= 0)
 						{
 							n.v1 = index.vertex;
@@ -305,14 +304,14 @@ namespace polyfem
 						n.nodes << V(nodes_ids[9], 0), V(nodes_ids[9], 1);
 					}
 				}
-				//P4
+				// P4
 				else if (nodes_ids.size() == 15)
 				{
 					orders_(f) = 4;
 					assert(false);
 					// unsupported P4 for geometry, need meshes for testing
 				}
-				//unsupported
+				// unsupported
 				else
 				{
 					assert(false);
@@ -362,13 +361,13 @@ namespace polyfem
 				}
 
 				assert(orders_(index.face) == 3);
-				//unsupported P4 for geometry
+				// unsupported P4 for geometry
 				const auto &n = face_nodes_[index.face];
 				return n.nodes.row(0);
 			}
 			else if (is_cube(index.face))
 			{
-				//supports only blilinear quads
+				// supports only blilinear quads
 				assert(orders_.size() <= 0 || orders_(index.face) == 1);
 
 				const auto v1 = point(index.vertex);
@@ -583,7 +582,7 @@ namespace polyfem
 			RowVectorNd min_corner, max_corner;
 			bounding_box(min_corner, max_corner);
 
-			//implement me properly
+			// implement me properly
 			for (int e = 0; e < n_edges(); ++e)
 			{
 				if (!is_boundary_edge(e))
@@ -610,7 +609,7 @@ namespace polyfem
 			boundary_ids_.resize(n_edges());
 			std::fill(boundary_ids_.begin(), boundary_ids_.end(), -1);
 
-			//implement me properly
+			// implement me properly
 			for (int e = 0; e < n_edges(); ++e)
 			{
 				if (!is_boundary_edge(e))
