+++ src/polyfem/utils/MatrixUtils.hpp
@@ -36,31 +36,6 @@ namespace polyfem
 			return I;
 		}
 
-		/// Reads a matrix from a file. Determines the file format based on the path's extension.
-		template <typename T>
-		bool read_matrix(const std::string &path, Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> &mat);
-
-		/// Writes a matrix to a file. Determines the file format based on the path's extension.
-		template <typename Mat>
-		bool write_matrix(const std::string &path, const Mat &mat);
-
-		template <typename T>
-		bool read_matrix_ascii(const std::string &path, Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> &mat);
-
-		template <typename Mat>
-		bool write_matrix_ascii(const std::string &path, const Mat &mat);
-
-		template <typename T>
-		bool read_matrix_binary(const std::string &path, Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> &mat);
-
-		template <typename Mat>
-		bool write_matrix_binary(const std::string &path, const Mat &mat);
-
-		bool write_sparse_matrix_csv(const std::string &path, const Eigen::SparseMatrix<double> &mat);
-
-		template <typename T>
-		bool import_matrix(const std::string &path, const json &import, Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> &mat);
-
 		class SpareMatrixCache
 		{
 		public:
